Fixing current build issues on Windows 64 bits with MSVC 12/2013

+++ src/ast.cpp
@@ -1828,7 +1828,7 @@ namespace Sass {
     {
       // do we have have too much precision?
       if (pos_fract < precision + pos_point)
-      { precision = pos_fract - pos_point; }
+      { precision = (int)(pos_fract - pos_point); }
       // round value again
       ss.precision(precision);
       ss << fixed << value_;
